Create parlist_to_json
This script prepares a JSON format from an excel file where a list of parameters in it. The created JSON format is used in the .cfg files of the embedded system. This automation is shortening the man/day's time and preventing wrong entries while the engineer is busy with adding parameters one by one.